Resolve conflict after rebase with master
