+++ app/db/schema.js
@@ -20,7 +20,7 @@ import schemaV16 from './schemas/schemaV16';
 import schemaV17 from './schemas/schemaV17';
 import schemaV19 from './schemas/v19/schemaV19';
 import schemaV20 from './schemas/v20/schemaV20';
-import schemaV21 from './schemas/schemaV21';
+import schemaV21 from './schemas/v21/schemaV21';
 import schemaV22 from './schemas/schemaV22';
 
 const schemas = [
