+++ app/db/schemas/v21/participant_migration_service.js
@@ -1,0 +1,16 @@
+const participantMigrationService = (() => {
+  return {
+    addCountable
+  }
+
+  function addCountable(oldRealm, newRealm) {
+    const oldParticipants = oldRealm.objects('Participant');
+    const newParticipants = newRealm.objects('Participant');
+
+    for (let i = 0; i < oldParticipants.length; i ++) {
+      newParticipants[i].countable = !oldParticipants[i].countable ? true : oldParticipants[i].countable;
+    }
+  }
+})();
+
+export default participantMigrationService;
